requirement.txt
absl-py==2.0.0
addict==2.4.0
adversarial-robustness-toolbox==1.16.0
annotated-types==0.7.0
anyio==4.2.0
anykeystore==0.2
apex==0.9.10.dev0
art==6.1
art-python==0.0.5
astor==0.8.1
asttokens==2.4.1
astunparse==1.6.3
attrdict==2.0.1
Babel==2.14.0
backcall==0.2.0
bce-python-sdk==0.8.98
beautifulsoup4==4.12.2
blinker==1.7.0
cachetools==5.3.1
certifi==2023.7.22
cffi==1.16.0
chardet==4.0.0
charset-normalizer==3.2.0
click==8.1.7
cloudpickle==3.0.0
clr-loader==0.2.6
coloredlogs==15.0.1
contourpy==1.1.0
coverage==7.3.2
cryptacular==1.6.2
cssselect==1.2.0
cssutils==2.9.0
curves==1.3.0
cycler==0.11.0
Cython==3.0.7
decorator==4.4.2
defusedxml==0.7.1
dnspython==2.6.1
easyocr==1.7.1
email_validator==2.2.0
et-xmlfile==1.1.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.111.1
fastapi-cli==0.0.4
filelock==3.15.4
fire==0.5.0
Flask==3.0.0
flask-babel==4.0.0
flatbuffers==23.5.26
fonttools==4.42.1
fsspec==2024.6.1
future==0.18.3
fvcore==0.1.5.post20221221
gast==0.3.3
gitdb==4.0.10
GitPython==3.1.36
google-auth==2.23.1
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
graphviz==0.20.1
greenlet==3.0.1
grpcio==1.59.2
h11==0.14.0
h5py==2.10.0
httpcore==1.0.2
httptools==0.6.1
httpx==0.26.0
huggingface-hub==0.24.5
humanfriendly==10.0
hupper==1.12
idna==2.10
imageio==2.31.6
imageio-ffmpeg==0.4.9
imgaug==0.4.0
importlib-metadata==6.8.0
importlib-resources==6.0.1
iopath==0.1.10
ipython==8.12.3
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
keras==2.13.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
lightning-utilities==0.10.0
lime==0.2.0.1
lmdb==1.4.1
lpips==0.1.4
lxml==4.9.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mmcv==2.0.0
mmcv-full==1.7.1
mmengine==0.8.5
moviepy==1.0.3
mpmath==1.3.0
networkx==3.1
ninja==1.11.1.1
numpy==1.24.1
nvidia-cublas-cu12==12.3.2.9
nvidia-cuda-cupti-cu12==12.3.52
nvidia-cuda-nvcc-cu12==12.3.52
nvidia-cuda-nvrtc-cu12==12.3.52
nvidia-cuda-runtime-cu12==12.3.52
nvidia-cudnn-cu12==8.9.6.50
nvidia-cufft-cu12==11.0.11.19
nvidia-curand-cu12==10.3.4.52
nvidia-cusolver-cu12==11.5.3.52
nvidia-cusparse-cu12==12.1.3.153
nvidia-dali-cuda110==1.23.0
nvidia-dali-nvtf-plugin==1.23.0+nv23.3
nvidia-horovod==0.27.0+nv23.3
nvidia-nccl-cu12==2.19.3
nvidia-nvjitlink-cu12==12.3.52
nvidia-pyindex==1.0.9
nvidia-tensorflow==1.15.5+nv23.3
oauthlib==3.2.2
onnxruntime==1.16.0
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
opencv-python-headless==4.8.0.76
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.2
paddleocr==2.7.0.3
paddlepaddle-gpu==2.5.2
pandas==2.0.3
parso==0.8.4
PasteDeploy==3.1.0
pbkdf2==1.3
pdf2docx==0.5.6
pexpect==4.9.0
pickleshare==0.7.5
Pillow==10.0.1
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==3.11.0
portalocker==2.8.2
premailer==3.10.0
proglog==0.1.10
progress==1.6
prompt_toolkit==3.0.46
protobuf==3.20.3
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyclipper==1.3.0.post5
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.8.2
pydantic_core==2.20.1
Pygments==2.16.1
PyMuPDF==1.20.2
pyparsing==3.0.9
pyramid==2.0.2
pyramid-mailer==0.15.1
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==1.1.0
python-dotenv==1.0.1
python-multipart==0.0.9
python3-openid==3.2.0
pythonnet==3.0.1
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0.1
rapidfuzz==3.5.2
rarfile==4.1
regex==2024.7.24
repoze.sendmail==4.4.1
requests==2.32.3
requests-oauthlib==1.3.1
rich==13.6.0
rsa==4.9
safetensors==0.4.3
scikit-image==0.21.0
scikit-learn==1.1.3
scipy==1.10.1
seaborn==0.12.2
shapely==2.0.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
stack-data==0.6.3
starlette==0.37.2
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow-estimator==2.13.0
tensorflow-hub==0.14.0
tensorflow-io-gcs-filesystem==0.34.0
tensorrt==8.6.1.post1
termcolor==2.3.0
thop==0.1.1.post2209072238
threadpoolctl==3.2.0
tifffile==2023.7.10
tokenizers==0.19.1
tomli==2.0.1
torch==1.13.1+cu116
torchattacks==3.5.1
torchaudio==0.13.1+cu116
torchmetrics==1.2.1
torchsummary==1.5.1
torchvision==0.14.1+cu116
torchviz==0.0.2
tqdm==4.66.1
traitlets==5.14.3
transaction==4.0
transformers==4.43.3
translationstring==1.4
typer==0.12.3
typing_extensions==4.5.0
tzdata==2023.3
ultralytics==8.0.178
urllib3==2.2.2
uvicorn==0.30.4
uvloop==0.19.0
velruse==1.1.1
venusian==3.1.0
visualdl==2.5.3
watchfiles==0.22.0
wcwidth==0.2.13
WebOb==1.8.7
websockets==12.0
Werkzeug==3.0.1
wrapt==1.15.0
WTForms==3.1.1
wtforms-recaptcha==0.3.2
yacs==0.1.8
yapf==0.40.2
zipp==3.16.2
zope.deprecation==5.0
zope.interface==6.1
zope.sqlalchemy==3.1
